docs: Results for all implementations added
The asl_recognizer html file and ipynb notebook 
contain now a complete table results  from running 
the recognizer using all the possible combinations 
of features and model selectors. 

To comply with instructions, the code for all the 
combinations was not included in these files. 
However the full implementation can be seen in 
the file:
asl_recognizer_all_combinations_when_selecting_models.ipynb